refactor (Car, CarName) : 이름 검증 기능 분리
- Car에 움직임 기능 등이 추가되면 Car에 많은 역할이 부과될 것을 우려하여 검증 기능 분리
